+++ java/com/turbovnc/rfb/CConnection.java
@@ -1,6 +1,6 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  * Copyright (C) 2011-2012 Brian P. Hinz
- * Copyright (C) 2012, 2014 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2012, 2014, 2016 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -219,7 +219,10 @@ public abstract class CConnection extends CMsgHandler {
   private void processSecurityResultMsg() {
     vlog.debug("processing security result message");
     int result;
-    if (cp.beforeVersion(3, 8) && csecurity.getType() == Security.secTypeNone) {
+    if (cp.beforeVersion(3, 8) &&
+        (csecurity.getType() == Security.secTypeNone ||
+         (csecurity instanceof CSecurityRFBTLS &&
+          csecurity.getType() == Security.secTypeTLSNone))) {
       result = Security.secResultOK;
     } else {
       if (!is.checkNoWait(1)) return;